refactored ses email service
